# Analyze and optimize code performance