… OPTIMIZATION_SUMMARY.md
… PERFORMANCE_ANALYSIS.md
… cmd/search_optimized_bench_test.go
… internal/client/client_bench_test.go
… internal/client/client_optimized_bench_test.go